make every work from TourDetails
